requirements.txt
flask>=1.1.2,<1.2
numpy>=1.18,<1.19
# opencv-contrib-python==4.1.0.25
Pillow>=7.2.0,<7.3
requests>=2.24.0,<2.25
scikit-image>=0.17.2,<0.18
scipy==1.4.1
